fix: rename separator
